backend/requirements.txt
langchain==0.0.137
openai
tiktoken
fastapi==0.95.0
uvicorn[standard]==0.21.1
hypercorn
gunicorn
pandas
weaviate-client
tqdm
python-dotenv